feat: 그래프 컴포넌트로 분리
