+++ src/app/_component/power/UsageGraph.tsx
@@ -1,5 +1,9 @@
 import React from "react";
+import Graph from "./Graph"; // Adjust the import path as necessary
+import { powerGraph } from "@/mock/powerGraph";
 
-export default function UsageGraph() {
-  return <>UsageGraph</>;
-}
+const UsageGraph = () => {
+  return <Graph data={powerGraph} isBillGraph={false} />;
+};
+
+export default UsageGraph;
